// packages/table/src/interface.ts
export type Key = React.Key
export type FixedType = 'left' | 'right' | boolean
export type DefaultRecordType = Record<string, any>
export type TableLayout = 'auto' | 'fixed'

// ==================== Row =====================
export type RowClassName<RecordType> = (
  record: RecordType,
  index: number,
  indent: number,
) => string

// =================== Column ===================
export interface CellType<RecordType> {
  key?: Key
  className?: string
  style?: React.CSSProperties
  children?: React.ReactNode
  column?: ColumnsType<RecordType>[number]
  colSpan?: number
  rowSpan?: number

  /** Only used for table header */
  hasSubColumns?: boolean
  colStart?: number
  colEnd?: number
}

export interface RenderedCell<RecordType> {
  props?: CellType<RecordType>
  children?: React.ReactNode
}

export type DataIndex = string | number | readonly (string | number)[]

export type CellEllipsisType = { showTitle?: boolean } | boolean

interface ColumnSharedType<RecordType> {
  title?: React.ReactNode
  key?: Key
  className?: string
  fixed?: FixedType
  onHeaderCell?: GetComponentProps<ColumnsType<RecordType>[number]>
  ellipsis?: CellEllipsisType
  align?: AlignType
}

export interface ColumnGroupType<RecordType> extends ColumnSharedType<RecordType> {
  children: ColumnsType<RecordType>
}

export type AlignType = 'left' | 'center' | 'right'

export interface ColumnType<RecordType> extends ColumnSharedType<RecordType> {
  colSpan?: number
  dataIndex?: DataIndex
  render?: (
    value: any,
    record: RecordType,
    index: number,
  ) => React.ReactNode | RenderedCell<RecordType>
  shouldCellUpdate?: (record: RecordType, prevRecord: RecordType) => boolean
  rowSpan?: number
  width?: number | string
  onCell?: GetComponentProps<RecordType>
  /** @deprecated Please use `onCell` instead */
  onCellClick?: (record: RecordType, e: React.MouseEvent<HTMLElement>) => void
}

export type ColumnsType<RecordType = unknown> = readonly (
  | ColumnGroupType<RecordType>
  | ColumnType<RecordType>
)[]

export type GetRowKey<RecordType> = (record: RecordType, index?: number) => Key

// ================= Fix Column =================
export interface StickyOffsets {
  left: readonly number[]
  right: readonly number[]
  isSticky?: boolean
}

// ================= Customized =================
export type GetComponentProps<DataType> = (
  data: DataType,
  index?: number,
) => React.HTMLAttributes<any> | React.TdHTMLAttributes<any>
// ...
export interface RenderExpandIconProps<RecordType> {
  prefixCls: string
  expanded: boolean
  record: RecordType
  expandable: boolean
  onExpand: TriggerEventHandler<RecordType>
}

export type RenderExpandIcon<RecordType> = (
  props: RenderExpandIconProps<RecordType>,
) => React.ReactNode
// ...
export type TriggerEventHandler<RecordType> = (
  record: RecordType,
  event: React.MouseEvent<HTMLElement>,
) => void